create generic functin to handle json urls
